Added support for wheel mice - the wn library now reports up to ten buttons, and there are new X resources to map button numbers to scrolling of the source window and/or display area.

+++ lib/libx11wn/wn_priv.h
@@ -160,7 +160,7 @@ typedef struct bwinst {
 	cursor_t bw_cid;
 	short bw_inmode;
 	short bw_upd_level;
-	short bw_winch_event;
+	unsigned long bw_winch_event;
 	short bw_had_lost_selection_event;
 	short bw_mouse_x;
 	short bw_mouse_y;
